requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --config=pyproject.toml requirements/requirements.in
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.0
    # via celery
brotli==1.1.0
    # via whitenoise
celery[redis]==5.4.0
    # via
    #   -r requirements/requirements.in
    #   django-celery-beat
celery-progress==0.3
    # via -r requirements/requirements.in
certifi==2024.6.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.3
    # via django-celery-beat
cryptography==42.0.7
    # via pyjwt
dj-rest-auth==6.0.0
    # via -r requirements/requirements.in
dj-stripe==2.8.4
    # via -r requirements/requirements.in
django==5.0.6
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   django-celery-beat
    #   django-cors-headers
    #   django-otp
    #   django-timezone-field
    #   django-waffle
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth[mfa,socialaccount]==0.63.3
    # via -r requirements/requirements.in
django-anymail[mailgun]==10.3
    # via -r requirements/requirements.in
django-celery-beat==2.6.0
    # via -r requirements/requirements.in
django-cors-headers==4.3.1
    # via -r requirements/requirements.in
django-environ==0.11.2
    # via -r requirements/requirements.in
django-otp==1.5.0
    # via -r requirements/requirements.in
django-timezone-field==6.1.0
    # via django-celery-beat
django-waffle==4.1.0
    # via -r requirements/requirements.in
djangorestframework==3.15.1
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-api-key==3.0.0
    # via -r requirements/requirements.in
djangorestframework-simplejwt==5.3.1
    # via -r requirements/requirements.in
docutils==0.21.2
    # via -r requirements/requirements.in
drf-spectacular==0.27.2
    # via -r requirements/requirements.in
idna==3.7
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.22.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.3.7
    # via celery
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.0
    # via djangorestframework-api-key
prompt-toolkit==3.0.45
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements/requirements.in
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
pypng==0.20220715.0
    # via qrcode
python-crontab==3.1.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   python-crontab
pyyaml==6.0.1
    # via drf-spectacular
qrcode==7.4.2
    # via django-allauth
redis==5.0.4
    # via celery
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
requests-oauthlib==2.0.0
    # via django-allauth
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
sqlparse==0.5.0
    # via django
stripe==4.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
typing-extensions==4.12.1
    # via qrcode
tzdata==2024.1
    # via
    #   celery
    #   django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.1
    # via
    #   django-anymail
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise[brotli]==6.6.0
    # via -r requirements/requirements.in
